refactor: 회원가입/사이드바/설정 화면 로직 및 UI 구조 분리
